Remove need for manual setup step
Instead of needing to run `.setup-better-dictionaries`,
now we run `./setup.sh` before each dict import.

+++ setup.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 # Make a copy of the original 'factory-version' workflow
-cp info.plist info.plist.orig
+if [ ! -f info.plist.orig ]; then
+  cp info.plist info.plist.orig
+fi
 
 chmod +x ./python
 chmod +x ./alfred-dict-server
